Refactor emergency contact phone links
